requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
mlflow
scikit-learn
dill
pyaml
dagshub
fastapi
uvicorn



##-e .